+++ src/main/java/lotto/exception/NotDuplicatedNumberException.java
@@ -1,4 +1,4 @@
-package lotto.domain.exception;
+package lotto.exception;
 
 public class NotDuplicatedNumberException extends RuntimeException{
     private static final String MESSAGE = "중복되는 로또 숫자를 입력하실 수 없습니다.";
