+++ src/main/java/lotto/lotto/LottoAutomaticTickets.java
@@ -1,0 +1,4 @@
+package lotto.lotto;
+
+public class LottoAutomaticTickets {
+}
