+++ src/main/java/lotto/lotto/LottoManualTicket.java
@@ -1,12 +1,11 @@
-package lotto.domain.lotto;
+package lotto.lotto;
 
 import java.util.List;
 
-public class LottoAutomaticTicket extends Lotto{
-
+public class LottoManualTicket extends Lotto{
     private final List<Integer> numbers;
 
-    public LottoAutomaticTicket(List<Integer> numbers) {
+    public LottoManualTicket(List<Integer> numbers) {
         super(numbers);
         this.numbers = numbers;
     }
