+++ src/main/java/lotto/lotto/LottoManualTickets.java
@@ -1,0 +1,4 @@
+package lotto.lotto;
+
+public class LottoManualTickets {
+}
