+++ src/main/java/lotto/validation/PurchaseAmountValidation.java
@@ -1,4 +1,4 @@
-package lotto.domain.validation;
+package lotto.validation;
 
 public class PurchaseAmountValidation {
     private static final int MIN_PURCHASE_AMOUNT = 0;
