+++ src/test/java/lotto/PurchaseAmountTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
